requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
attrs==24.2.0
bitarray==2.9.2
certifi==2024.8.30
charset-normalizer==3.3.2
ckzg==2.0.1
cytoolz==0.12.3
eth-account==0.13.3
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==2.1.0
eth-typing==5.0.0
eth-utils==5.0.0
eth_abi==5.1.0
frozenlist==1.4.1
hexbytes==1.2.1
idna==3.10
multidict==6.1.0
parsimonious==0.10.0
pycryptodome==3.20.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
pyunormalize==16.0.0
regex==2024.9.11
requests==2.32.3
rlp==4.0.1
toolz==0.12.1
types-requests==2.32.0.20240914
typing_extensions==4.12.2
urllib3==2.2.3
w3multicall==0.3.1
web3==7.2.0
websockets==13.0.1
yarl==1.11.1
